refactor(nms): Consolidate LteNetworkContext
