+++ nms/app/views/network/__tests__/NetworkTest.tsx
@@ -20,10 +20,6 @@ import FEGNetworkDashboard from '../FEGNetworkDashboard';
 import GatewayContext, {
   GatewayContextType,
 } from '../../../components/context/GatewayContext';
-import LteNetworkContext, {
-  LteNetworkContextType,
-  UpdateNetworkContextProps,
-} from '../../../components/context/LteNetworkContext';
 import MuiStylesThemeProvider from '@material-ui/styles/ThemeProvider';
 import NetworkDashboard from '../NetworkDashboard';
 import PolicyContext, {
@@ -35,18 +31,17 @@ import SubscriberContext, {
 } from '../../../components/context/SubscriberContext';
 import axiosMock from 'axios';
 import defaultTheme from '../../../theme/default';
+import {LteNetworkContextProvider} from '../../../components/context/LteNetworkContext';
 
 import {MemoryRouter, Route, Routes} from 'react-router-dom';
 import {MuiThemeProvider} from '@material-ui/core/styles';
-import {
-  UpdateNetworkProps,
-  UpdateNetworkState,
-} from '../../../state/lte/NetworkState';
 import {fireEvent, render, wait} from '@testing-library/react';
 import {forbiddenNetworkTypes} from '../../subscriber/SubscriberUtils';
 
 import MagmaAPI from '../../../api/MagmaAPI';
+import NetworkContext from '../../../components/context/NetworkContext';
 import axios from 'axios';
+import {LTE} from '../../../../shared/types/network';
 import {mockAPI} from '../../../util/TestUtils';
 import {useEnqueueSnackbar} from '../../../hooks/useSnackbar';
 import type {FegNetwork, NetworkEpcConfigs} from '../../../../generated';
@@ -262,6 +257,8 @@ describe('<NetworkDashboard />', () => {
     },
   };
 
+  let lteNetworkIdGetMock: jest.SpyInstance;
+
   beforeEach(() => {
     (useEnqueueSnackbar as jest.MockedFunction<
       typeof useEnqueueSnackbar
@@ -270,7 +267,15 @@ describe('<NetworkDashboard />', () => {
     (axiosMock as jest.Mocked<typeof axios>).post.mockImplementation(() =>
       Promise.resolve({data: {success: true}}),
     );
-    mockAPI(MagmaAPI.lteNetworks, 'lteNetworkIdGet');
+
+    lteNetworkIdGetMock = mockAPI(MagmaAPI.lteNetworks, 'lteNetworkIdGet', {
+      ...testNetwork,
+      cellular: {
+        epc: epc,
+        ran: ran,
+      },
+    });
+
     mockAPI(MagmaAPI.lteNetworks, 'lteNetworkIdPut');
 
     mockAPI(MagmaAPI.lteNetworks, 'lteNetworkIdCellularEpcPut');
@@ -320,46 +325,34 @@ describe('<NetworkDashboard />', () => {
       refetchSessionState: () => {},
     } as SubscriberContextType;
 
-    const networkCtx = {
-      state: {
-        ...testNetwork,
-
-        cellular: {
-          epc: epc,
-          ran: ran,
-        },
-      },
-      updateNetworks: async (props: UpdateNetworkContextProps) => {
-        return UpdateNetworkState({
-          setLteNetwork: () => {},
-          refreshState: testNetwork.id === props.networkId,
-          ...props,
-        } as UpdateNetworkProps); // TODO[TS-migration] Broken LteNetworkContext type
-      },
-    } as LteNetworkContextType;
-
     return (
       <MemoryRouter initialEntries={['/nms/test/network']} initialIndex={0}>
         <MuiThemeProvider theme={defaultTheme}>
           <MuiStylesThemeProvider theme={defaultTheme}>
-            <LteNetworkContext.Provider value={networkCtx}>
-              <PolicyContext.Provider value={policyCtx}>
-                <ApnContext.Provider value={apnCtx}>
-                  <GatewayContext.Provider value={gatewayCtx}>
-                    <EnodebContext.Provider value={enodebCtx}>
-                      <SubscriberContext.Provider value={subscriberCtx}>
-                        <Routes>
-                          <Route
-                            path="/nms/:networkId/network/*"
-                            element={<NetworkDashboard />}
-                          />
-                        </Routes>
-                      </SubscriberContext.Provider>
-                    </EnodebContext.Provider>
-                  </GatewayContext.Provider>
-                </ApnContext.Provider>
-              </PolicyContext.Provider>
-            </LteNetworkContext.Provider>
+            <NetworkContext.Provider
+              value={{
+                networkId: testNetwork.id,
+                networkType: LTE,
+              }}>
+              <LteNetworkContextProvider networkId={testNetwork.id}>
+                <PolicyContext.Provider value={policyCtx}>
+                  <ApnContext.Provider value={apnCtx}>
+                    <GatewayContext.Provider value={gatewayCtx}>
+                      <EnodebContext.Provider value={enodebCtx}>
+                        <SubscriberContext.Provider value={subscriberCtx}>
+                          <Routes>
+                            <Route
+                              path="/nms/:networkId/network/*"
+                              element={<NetworkDashboard />}
+                            />
+                          </Routes>
+                        </SubscriberContext.Provider>
+                      </EnodebContext.Provider>
+                    </GatewayContext.Provider>
+                  </ApnContext.Provider>
+                </PolicyContext.Provider>
+              </LteNetworkContextProvider>
+            </NetworkContext.Provider>
           </MuiStylesThemeProvider>
         </MuiThemeProvider>
       </MemoryRouter>
@@ -423,6 +416,8 @@ describe('<NetworkDashboard />', () => {
   it('Verify Network Add', async () => {
     const {getByTestId, getByText, queryByTestId} = render(<Wrapper />);
     await wait();
+    lteNetworkIdGetMock.mockClear();
+
     expect(queryByTestId('editDialog')).toBeNull();
 
     fireEvent.click(getByText('Add Network'));
@@ -586,12 +581,14 @@ describe('<NetworkDashboard />', () => {
       },
       networkId: 'testNetworkID',
     });
-    expect(MagmaAPI.lteNetworks.lteNetworkIdGet).toHaveBeenCalledTimes(0);
+    expect(lteNetworkIdGetMock).toHaveBeenCalledTimes(0);
   });
 
   it('Verify Network Edit Info', async () => {
     const {getByTestId, getByText, queryByTestId} = render(<Wrapper />);
     await wait();
+    lteNetworkIdGetMock.mockClear();
+
     expect(queryByTestId('editDialog')).toBeNull();
 
     fireEvent.click(getByTestId('infoEditButton'));
@@ -638,12 +635,14 @@ describe('<NetworkDashboard />', () => {
         },
       },
     });
-    expect(MagmaAPI.lteNetworks.lteNetworkIdGet).toHaveBeenCalledTimes(1);
+    expect(lteNetworkIdGetMock).toHaveBeenCalledTimes(1);
   });
 
   it('Verify Network Edit EPC', async () => {
     const {getByTestId, getByText, queryByTestId} = render(<Wrapper />);
     await wait();
+    lteNetworkIdGetMock.mockClear();
+
     expect(queryByTestId('editDialog')).toBeNull();
 
     fireEvent.click(getByTestId('epcEditButton'));
@@ -668,12 +667,14 @@ describe('<NetworkDashboard />', () => {
       config: {...epc, mnc: '03'},
       networkId: 'test_network',
     });
-    expect(MagmaAPI.lteNetworks.lteNetworkIdGet).toHaveBeenCalledTimes(1);
+    expect(lteNetworkIdGetMock).toHaveBeenCalledTimes(1);
   });
 
   it('Verify Network Edit Ran', async () => {
     const {getByTestId, getByText, queryByTestId} = render(<Wrapper />);
     await wait();
+    lteNetworkIdGetMock.mockClear();
+
     expect(queryByTestId('editDialog')).toBeNull();
 
     fireEvent.click(getByTestId('ranEditButton'));
@@ -704,7 +705,7 @@ describe('<NetworkDashboard />', () => {
       },
       networkId: 'test_network',
     });
-    expect(MagmaAPI.lteNetworks.lteNetworkIdGet).toHaveBeenCalledTimes(1);
+    expect(lteNetworkIdGetMock).toHaveBeenCalledTimes(1);
   });
 });
